requirements.txt
python-slugify
flask
flask-sqlalchemy
flask-security
py-bcrypt
flask-markdown
pygments
flask-admin
Flask-Script
WTForms-Alchemy
SQLAlchemy-Wrapper
psycopg2